extract helpers from http_server.py to http.py
