+++ wazo_dird/http_server.py
@@ -1,26 +1,20 @@
 # Copyright 2016-2022 The Wazo Authors  (see the AUTHORS file)
 # SPDX-License-Identifier: GPL-3.0-or-later
+
 import logging
 import os
-import time
 
 from datetime import timedelta
-from functools import wraps
 
-from requests import HTTPError
 from cheroot import wsgi
 from flask import Flask
-from flask import request
 from flask_restful import Api
-from flask_restful import Resource
 from flask_cors import CORS
-from wazo_auth_client import Client as AuthClient
 from werkzeug.contrib.fixers import ProxyFix
-from xivo.auth_verifier import AuthVerifier
 from xivo import http_helpers
-from xivo import mallow_helpers
-from xivo import rest_api_helpers
 from xivo.http_helpers import ReverseProxied
+
+from .http import AuthResource, auth_verifier
 
 logger = logging.getLogger(__name__)
 
@@ -29,7 +23,6 @@ TEMPLATE_FOLDER = 'plugins/templates'
 
 app = Flask('wazo-setupd')
 api = Api(app, prefix='/{}'.format(VERSION))
-auth_verifier = AuthVerifier()
 
 
 class CoreRestApi:
@@ -87,63 +80,3 @@ class CoreRestApi:
     def stop(self):
         if self.server:
             self.server.stop()
-
-
-def handle_api_exception(func):
-    @wraps(func)
-    def wrapper(*args, **kwargs):
-        try:
-            return func(*args, **kwargs)
-        except rest_api_helpers.APIException as error:
-            response = {
-                'reason': [error.message],
-                'timestamp': time.time(),
-                'status_code': error.status_code,
-            }
-            logger.error('%s: %s', error.message, error.details)
-            return response, error.status_code
-
-    return wrapper
-
-
-class LegacyErrorCatchingResource(Resource):
-    method_decorators = [handle_api_exception] + Resource.method_decorators
-
-
-class LegacyAuthResource(LegacyErrorCatchingResource):
-    method_decorators = [
-        auth_verifier.verify_token
-    ] + LegacyErrorCatchingResource.method_decorators
-
-
-class ErrorCatchingResource(Resource):
-    method_decorators = [
-        mallow_helpers.handle_validation_exception,
-        rest_api_helpers.handle_api_exception,
-    ] + Resource.method_decorators
-
-
-class AuthResource(ErrorCatchingResource):
-
-    method_decorators = [
-        auth_verifier.verify_token
-    ] + ErrorCatchingResource.method_decorators
-
-    def get_visible_tenants(self, tenant):
-        token = request.headers['X-Auth-Token']
-        auth_client = AuthClient(**self.auth_config)
-        auth_client.set_token(token)
-
-        try:
-            visible_tenants = auth_client.tenants.list(tenant_uuid=tenant)['items']
-        except HTTPError as e:
-            response = getattr(e, 'response', None)
-            status_code = getattr(response, 'status_code', None)
-            if status_code == 401:
-                logger.warning(
-                    'a user is doing multi-tenant queries without the tenant list ACL'
-                )
-                return [tenant]
-            raise
-
-        return [tenant['uuid'] for tenant in visible_tenants]
