+++ wazo_dird/plugins/backends/http.py
@@ -4,7 +4,7 @@
 from flask import request
 
 from wazo_dird.auth import required_acl
-from wazo_dird.http_server import LegacyAuthResource
+from wazo_dird.http import LegacyAuthResource
 
 from .schemas import ListSchema
