+++ wazo_dird/plugins/conference_backend/http.py
@@ -7,7 +7,7 @@ from xivo.tenant_flask_helpers import Tenant
 from wazo_dird import exception
 from wazo_dird.helpers import SourceItem, SourceList
 from wazo_dird.auth import required_acl
-from wazo_dird.http_server import AuthResource
+from wazo_dird.http import AuthResource
 from wazo_dird.plugin_helpers.confd_client_registry import registry
 
 from .schemas import (
