+++ wazo_dird/plugins/google_backend/http.py
@@ -6,7 +6,7 @@ import logging
 from flask import request
 from wazo_dird.auth import required_acl
 from wazo_dird.helpers import SourceItem, SourceList
-from wazo_dird.http_server import AuthResource
+from wazo_dird.http import AuthResource
 from xivo.tenant_flask_helpers import Tenant, token
 
 from .schemas import contact_list_schema, list_schema, source_list_schema, source_schema
