+++ wazo_dird/plugins/headers/http.py
@@ -8,7 +8,7 @@ from xivo.tenant_flask_helpers import Tenant
 from wazo_dird.exception import OldAPIException
 from wazo_dird.helpers import DisplayAwareResource
 from wazo_dird.auth import required_acl
-from wazo_dird.http_server import LegacyAuthResource
+from wazo_dird.http import LegacyAuthResource
 
 logger = logging.getLogger(__name__)
