+++ wazo_dird/plugins/profiles/http.py
@@ -7,7 +7,7 @@ from flask import request
 from xivo.tenant_flask_helpers import Tenant
 
 from wazo_dird.auth import required_acl
-from wazo_dird.http_server import AuthResource
+from wazo_dird.http import AuthResource
 
 from .schemas import list_schema, profile_list_schema, profile_schema
