+++ wazo_dird/plugins/sources/http.py
@@ -5,7 +5,7 @@ from flask import request
 from xivo.tenant_flask_helpers import Tenant
 
 from wazo_dird.auth import required_acl
-from wazo_dird.http_server import AuthResource
+from wazo_dird.http import AuthResource
 
 from .schemas import source_list_schema, list_schema
