+++ wazo_dird/plugins/status/http.py
@@ -2,7 +2,7 @@
 # SPDX-License-Identifier: GPL-3.0-or-later
 
 from wazo_dird.auth import required_acl
-from wazo_dird.http_server import AuthResource
+from wazo_dird.http import AuthResource
 
 
 class StatusResource(AuthResource):
